Move numpydoc to sklearn/externals and ignore externals directory in check
