+++ sklearn/externals/numpy_ext/docscrape.py
