+++ sklearn/externals/numpy_ext/docscrape_sphinx.py
