+++ sklearn/externals/numpy_ext/numpydoc.py
